Move BackupsTab out of ui.py
